backend/requirements.txt
Flask==3.1.1
flask-cors==5.0.1
flask-limiter==3.12
python-dotenv==1.1.0
requests==2.32.3
cachetools==5.5.2
gunicorn==23.0.0
pytest==8.4.0
